Create injectMidiChord(NotMidiEnableFunction).js
This Javascript when run will hyperlink the chords displayed on Worshiptogether.com song chords page to send Midi notes of the Chords. However this does not inject function script to enable Midi. This is done by extracting the "matchedChords" class html tag and "Suffix" class html tag and send those to a noteon function with these two parameters when the button is clicked. When the suffix is not available it sends a '0' so a Major Chord is played but when suffix 'm' is sent the middle note of the chord is reduced by 1 note to produce a minor chord.

However for displaying the chords on the webpage the innerText is extracted from the html class tag "chord-pro-note" which is an exact representation of the chords displayed